Added a function with an extra parameter to use the "addtional_owners" in twitter's API
The added parameter "additionalOwners" is corresponding to the optional
parameter "addtional_owners" in the twitter's document, which is a
comma-separated list of user IDs to set as additional owners allowed to
use the returned media_id in Tweets or Cards. Up to 100 additional
owners may be specified.

This parameter will be useful when someone wants to post a media tweet
by multiple twitter accounts at one time without having to uploading
the media for multiple times, which is a waste of networking.