Add support for Linux as a target.
This required guarding anything that uses the Apple-specific Accounts
and Social frameworks with #if configurations, removing any instances of
bridging coercions between NS-types and their value equivalents, and
implementing small wrapper APIs around API differences in NSScanner on
Linux. In the future, API parity will render at least the Scanner
changes obsolete.

This patch also reorganizes the tests so the test code is shared
transparently with the Linux target, and adds some tests to the query
parser.